+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - "Disable autofix" rules are now generated for all the vanilla ESLint rules.
 - [**BREAKING**] "Disable autofix" rules are no longer generated for rules that do not provide fixes.
 - Set [`sonarjs/no-clear-text-protocols`](https://sonarsource.github.io/rspec/#/rspec/S5332/javascript) rule severity to `warning` because it might produce many false positives.
+- [**BREAKING**] `ts` config: added `configTypeAware` sub-config to replace `filesTypeAware`, `ignoresTypeAware` and `overridesTypeAware` options. Unless explicitly specified, it will use the same `ignores` as the parent config.
+- [**BREAKING**] Enabled [`@typescript-eslint/consistent-return`](https://typescript-eslint.io/rules/consistent-return) rule in `ts` config and disabled the same base rule [`consistent-return`](https://eslint.org/docs/latest/rules/consistent-return).
 
 ## 0.6.2